backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
anthropic>=0.40,<1.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
